De-duplicate logic between testing.js and the rest of the source code
